moved intervention model to struct to seperate files
